week2/Web Scraping & HTML Cleaning/requirements.txt
requests>=2.28.0
trafilatura>=1.6.0
pytesseract>=0.3.10
selenium>=4.0.0
Pillow>=9.0.0
lxml>=4.9.0
